requirement.txt
# python>=3.7.0

numpy==1.23.4
pytroch==1.12.1
torchvison==0.13.1
monai==1.0.1
xlwt==1.3.0
SimpleITK==2.2.0
scipy==1.9.3
tqdm==4.64.1
matplotlib==3.6.1
sklearn

# if need gradcam
cv2